chore: move assert up
wip: remove barrel file

wip
