--- src/kairos/shared/kernel/assert/index.ts
@@ -1,2 +1,0 @@
-export { assertIsNonEmptyString, assertIsString } from './string-assertions';
-export { assertOneOf } from './value-assertions';
